Change RunnerProfitAndLoss.selection_id type to int
Previously RunnerProfitAndLoss.selection_id was declared to be a float, however the
API documentation says that it is a long:

https://api.developer.betfair.com/services/webapps/docs/x/SIA6#BettingTypeDefinitions-RunnerProfitAndLoss

This changes the declared data type in the model to int. int is used
over long as this maintains Python 3 compatability (in Python 2
int('<some number>') will return a long if the number is big enough
to require it).

+++ betfair/models.py
@@ -172,7 +172,7 @@ class MarketBook(BetfairModel):
 
 
 class RunnerProfitAndLoss(BetfairModel):
-    selection_id = Field(DataType(float))
+    selection_id = Field(DataType(int))
     if_win = Field(DataType(float))
     if_lose = Field(DataType(float))
